server/requirements.txt
# FastAPI 서버 의존성
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0

# DB
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.0

# Celery (비동기 워커)
celery[redis]==5.4.0

# 인증
pyjwt==2.9.0
httpx==0.27.0

# S3 호환 스토리지
boto3==1.35.0

# 설정
pydantic-settings==2.6.0

# 과학 연산 (DNA 분석)
numpy==2.1.0
scipy==1.14.0

# 테스트
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
